+++ src/agent/mod.rs
@@ -1,5 +1,7 @@
+mod provider;
 mod rt;
 mod spec;
 
+pub use provider::*;
 pub use rt::*;
 pub use spec::*;
