+++ src/lang_model/mod.rs
@@ -1,5 +1,5 @@
+mod provider;
 mod rt;
-mod spec;
 
+pub use provider::*;
 pub use rt::*;
-pub use spec::*;
