+++ src/tool/func.rs
