+++ src/tool/mod.rs
@@ -1,5 +1,9 @@
+mod func;
+mod provider;
 mod rt;
-mod spec;
+mod toolset;
 
+pub use func::*;
+pub use provider::*;
 pub use rt::*;
-pub use spec::*;
+pub use toolset::*;
