--- src/tool/rt/mod.rs
@@ -1,7 +1,0 @@
-mod func;
-mod runtime;
-mod toolset;
-
-pub use func::*;
-pub use runtime::*;
-pub use toolset::*;
